APIMF-3447/0: moved DataNodeParser/Emitter to amf-core so I could use it from amf-aml
