+++ amf-api-contract/shared/src/main/scala/amf/apicontract/internal/spec/common/parser/WebApiContext.scala
@@ -14,6 +14,7 @@ import amf.core.client.scala.model.document.{ExternalFragment, Fragment, Recursi
 import amf.core.client.scala.model.domain.extensions.CustomDomainProperty
 import amf.core.client.scala.model.domain.{AmfObject, Shape}
 import amf.core.client.scala.parse.document.{ParsedReference, ParserContext}
+import amf.core.internal.datanode.DataNodeParserContext
 import amf.core.internal.parser._
 import amf.core.internal.parser.domain.{Annotations, FragmentRef, SearchScope}
 import amf.core.internal.plugins.syntax.{SYamlAMFParserErrorHandler, SyamlAMFErrorHandler}
@@ -24,7 +25,6 @@ import amf.shapes.client.scala.model.domain.AnyShape
 import amf.shapes.internal.spec.common.parser.{SpecSyntax, YMapEntryLike}
 import amf.shapes.internal.spec.common.{JSONSchemaDraft4SchemaVersion, SchemaVersion}
 import amf.shapes.internal.spec.contexts.JsonSchemaRefGuide
-import amf.shapes.internal.spec.datanode.DataNodeParserContext
 import amf.shapes.internal.spec.jsonschema.ref.{AstFinder, AstIndex, AstIndexBuilder, JsonSchemaInference}
 import org.mulesoft.lexer.SourceLocation
 import org.yaml.model._
