+++ amf-shapes/shared/src/main/scala/amf/shapes/internal/spec/ShapeParserContext.scala
@@ -5,6 +5,7 @@ import amf.core.client.scala.config.ParsingOptions
 import amf.core.client.scala.errorhandling.AMFErrorHandler
 import amf.core.client.scala.model.domain.{AmfObject, Shape}
 import amf.core.client.scala.parse.document.{ErrorHandlingContext, UnresolvedComponents}
+import amf.core.internal.datanode.DataNodeParserContext
 import amf.core.internal.parser.Root
 import amf.core.internal.parser.domain.{Annotations, Declarations, FutureDeclarations, SearchScope}
 import amf.core.internal.plugins.syntax.SyamlAMFErrorHandler
@@ -14,7 +15,6 @@ import amf.shapes.internal.spec.RamlWebApiContextType.RamlWebApiContextType
 import amf.shapes.internal.spec.common.SchemaVersion
 import amf.shapes.internal.spec.common.parser.{SpecSyntax, YMapEntryLike}
 import amf.shapes.internal.spec.contexts.JsonSchemaRefGuide
-import amf.shapes.internal.spec.datanode.DataNodeParserContext
 import amf.shapes.internal.spec.raml.parser.external.RamlExternalTypesParser
 import amf.shapes.internal.spec.raml.parser.{DefaultType, RamlTypeParser, TypeInfo}
 import org.mulesoft.lexer.SourceLocation
