+++ amf-shapes/shared/src/main/scala/amf/shapes/internal/spec/common/emitter/EnumValuesEmitter.scala
@@ -2,6 +2,7 @@ package amf.shapes.internal.spec.common.emitter
 
 import amf.core.client.common.position.Position
 import amf.core.client.scala.model.domain.{AmfArray, DataNode}
+import amf.core.internal.datanode.DataNodeEmitter
 import amf.core.internal.parser.domain.Value
 import amf.core.internal.render.BaseEmitters.{pos, traverse}
 import amf.core.internal.render.SpecOrdering
