+++ amf-shapes/shared/src/main/scala/amf/shapes/internal/spec/common/emitter/PayloadEmitter.scala
@@ -2,6 +2,7 @@ package amf.shapes.internal.spec.common.emitter
 
 import amf.core.client.scala.errorhandling.AMFErrorHandler
 import amf.core.client.scala.model.domain.DataNode
+import amf.core.internal.datanode.DataNodeEmitter
 import amf.core.internal.render.SpecOrdering
 import org.yaml.model.YDocument
