+++ amf-shapes/shared/src/main/scala/amf/shapes/internal/spec/common/emitter/annotations/FacetsInstanceEmitter.scala
@@ -3,11 +3,12 @@ package amf.shapes.internal.spec.common.emitter.annotations
 import amf.core.client.common.position.Position
 import amf.core.client.scala.model.domain.Shape
 import amf.core.client.scala.model.domain.extensions.ShapeExtension
+import amf.core.internal.datanode.DataNodeEmitter
 import amf.core.internal.render.BaseEmitters.pos
 import amf.core.internal.render.SpecOrdering
 import amf.core.internal.render.emitters.EntryEmitter
 import amf.core.internal.utils.AmfStrings
-import amf.shapes.internal.spec.common.emitter.{DataNodeEmitter, ShapeEmitterContext}
+import amf.shapes.internal.spec.common.emitter.ShapeEmitterContext
 import org.yaml.model.YDocument.EntryBuilder
 
 case class FacetsEmitter(element: Shape, ordering: SpecOrdering)(implicit spec: ShapeEmitterContext) {
