+++ amf-shapes/shared/src/main/scala/amf/shapes/internal/spec/common/parser/CommonEnumParser.scala
@@ -2,8 +2,8 @@ package amf.shapes.internal.spec.common.parser
 
 import amf.core.client.scala.model.domain.DataNode
 import amf.core.client.scala.parse.document.ErrorHandlingContext
+import amf.core.internal.datanode.{DataNodeParser, DataNodeParserContext, ScalarNodeParser}
 import amf.core.internal.utils.IdCounter
-import amf.shapes.internal.spec.datanode.{DataNodeParser, DataNodeParserContext, ScalarNodeParser}
 import org.yaml.model.{IllegalTypeHandler, YNode}
 
 object EnumParsing {
@@ -15,7 +15,6 @@ object CommonEnumParser {
 
   def apply(parentId: String, enumType: String = EnumParsing.UNKNOWN_ENUM)(
       implicit ctx: ErrorHandlingContext with DataNodeParserContext with IllegalTypeHandler): YNode => DataNode = {
-    val enumParentId = s"$parentId/enum"
     enumType match {
       case EnumParsing.SCALAR_ENUM => ScalarNodeParser().parse
       case _                       => DataNodeParser.parse(idCounter = new IdCounter())
