+++ amf-shapes/shared/src/main/scala/amf/shapes/internal/spec/datanode/AbstractVariables.scala
@@ -1,18 +1,22 @@
 package amf.shapes.internal.spec.datanode
 
 import amf.core.client.scala.model.domain.AmfScalar
+import amf.core.internal.datanode.OnScalarParseHook
 import amf.core.internal.transform.VariableReplacer.VariableRegex
 import amf.core.internal.parser.domain.{Annotations, ScalarNode}
 import org.yaml.model.YNode
 
 import scala.collection.mutable
 
-case class AbstractVariables() {
+case class AbstractVariables() extends OnScalarParseHook {
   private val variables: mutable.Map[String, Annotations] = mutable.Map()
 
-  def parseVariables(node: YNode): this.type = parseVariables(ScalarNode(node))
+  override def onScalarParse(node: YNode): Unit      = parseVariables(node)
+  override def onScalarParse(node: ScalarNode): Unit = parseVariables(node)
 
-  def parseVariables(scalarNode: ScalarNode): this.type = {
+  private def parseVariables(node: YNode): this.type = parseVariables(ScalarNode(node))
+
+  private def parseVariables(scalarNode: ScalarNode): this.type = {
     VariableRegex
       .findAllMatchIn(scalarNode.text().toString)
       .foreach(m => variables.update(m.group(1), scalarNode.string().annotations))
