+++ amf-shapes/shared/src/main/scala/amf/shapes/internal/spec/jsonschema/ref/AstIndexBuilder.scala
@@ -1,11 +1,11 @@
 package amf.shapes.internal.spec.jsonschema.ref
 
 import amf.core.internal.utils.AliasCounter
+import amf.core.internal.validation.CoreValidations.ExceededMaxYamlReferences
 import amf.shapes.internal.spec.ShapeParserContext
-import amf.shapes.internal.spec.common.parser.YMapEntryLike
 import amf.shapes.internal.spec.common._
+import amf.shapes.internal.spec.common.parser.YMapEntryLike
 import amf.shapes.internal.spec.jsonschema.ref
-import amf.shapes.internal.validation.definitions.ShapeParserSideValidations.ExceededMaxYamlReferences
 import org.yaml.model._
 
 import java.net.URI
